+++ apps/explorer/lib/explorer/smart_contract/rust_verifier_interface.ex
@@ -1,0 +1,6 @@
+defmodule Explorer.SmartContract.RustVerifierInterface do
+  @moduledoc """
+    Adapter for contracts verification with https://github.com/blockscout/blockscout-rs/blob/main/smart-contract-verifier
+  """
+  use Explorer.SmartContract.RustVerifierInterfaceBehaviour
+end
